+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/modes/SpriteExportMode.java
@@ -29,6 +29,10 @@ public enum SpriteExportMode {
      */
     PNG,
     /**
+     * APNG - Animated Portable Network Graphics
+     */
+    APNG,
+    /**
      * GIF - Graphics Interchange Format
      */
     GIF,